Balacing
Edited SO for first pass at balancing
Added more Evil Points (30)
Adjusted EP cost for each entity
Added functionality to Remove button (only remove all, hid remove one button for now)

+++ Assets/Dossiers Personnels/Sarah/Script/S_O/SpikeStats_SO.cs
@@ -16,5 +16,5 @@ public class SpikeStats_SO : ScriptableObject
     [field: SerializeField]
     public float kockbackForce { get; private set; }
     [field: SerializeField]
-    public float cost { get; private set; }
+    public int cost { get; private set; }
 }
